iarap/requirements.txt
anyio==4.5.2
asgiref==3.8.1
backports.zoneinfo==0.2.1
cachecontrol==0.14.2
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
django-cors-headers==4.3.1
cryptography==45.0.6
django==4.2.23
djangorestframework==3.15.2
exceptiongroup==1.3.0
firebase-admin==6.9.0
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.3.0
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==21.2.0
h11==0.16.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.0.1
idna==3.10
msgpack==1.1.1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.1.4
python-dotenv==1.0.1
requests==2.32.4
rsa==4.9.1
sniffio==1.3.1
sqlparse==0.5.3
typing-extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.2.3
